- Themes support; - Geo location bug fixing;

+++ PhotoDB/Units/uGraphicUtils.pas
@@ -105,7 +105,7 @@ end;
 function MakeDarken(Color: TColor): TColor;
 begin
   Color := ColorToRGB(Color);
-  Result := RGB(Round(0.75 * GetRValue(Color)), Round(0.75 * GetGValue(Color)), Round(0.75 * GetBValue(Color)));
+  Result := RGB(Byte(Round(0.75 * GetRValue(Color))), Byte(Round(0.75 * GetGValue(Color))), Byte(Round(0.75 * GetBValue(Color))));
 end;
 
 function MixColors(Color1, Color2: TColor; Percent: Integer): TColor;
@@ -116,9 +116,9 @@ begin
   Color1 := ColorToRGB(Color1);
   Color2 := ColorToRGB(Color2);
   P := (Percent / 100);
-  R := Round(P * GetRValue(Color1) + (P - 1) * GetRValue(Color2));
-  G := Round(P * GetGValue(Color1) + (P - 1) * GetGValue(Color2));
-  B := Round(P * GetBValue(Color1) + (P - 1) * GetBValue(Color2));
+  R := Byte(Round(P * GetRValue(Color1) + (P - 1) * GetRValue(Color2)));
+  G := Byte(Round(P * GetGValue(Color1) + (P - 1) * GetGValue(Color2)));
+  B := Byte(Round(P * GetBValue(Color1) + (P - 1) * GetBValue(Color2)));
   Result := RGB(R, G, B);
 end;
